Pass 1 of resharper auto-changes

+++ src/Infrastructure.Network/Disk/ReddeerDirectory.cs
@@ -1,25 +1,20 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using Infrastructure.Network.Disk.Interfaces;
-
-namespace Infrastructure.Network.Disk
+﻿namespace Infrastructure.Network.Disk
 {
+    using System;
+    using System.Collections.Generic;
+    using System.IO;
+    using System.Linq;
+
+    using Infrastructure.Network.Disk.Interfaces;
+
     public class ReddeerDirectory : IReddeerDirectory
     {
         public bool Create(string path)
         {
-            if (string.IsNullOrWhiteSpace(path))
-            {
-                return false;
-            }
+            if (string.IsNullOrWhiteSpace(path)) return false;
 
             var invalidCharacters = Path.GetInvalidPathChars();
-            if (path.ToCharArray().Any(pa => invalidCharacters.Contains(pa)))
-            {
-                throw new ArgumentException();
-            }
+            if (path.ToCharArray().Any(pa => invalidCharacters.Contains(pa))) throw new ArgumentException();
 
             try
             {
@@ -35,23 +30,14 @@ namespace Infrastructure.Network.Disk
 
         public bool Delete(string path)
         {
-            if (string.IsNullOrWhiteSpace(path))
-            {
-                return false;
-            }
+            if (string.IsNullOrWhiteSpace(path)) return false;
 
             var invalidCharacters = Path.GetInvalidPathChars();
-            if (path.ToCharArray().Any(pa => invalidCharacters.Contains(pa)))
-            {
-                throw new ArgumentException();
-            }
+            if (path.ToCharArray().Any(pa => invalidCharacters.Contains(pa))) throw new ArgumentException();
 
             try
             {
-                if (Directory.Exists(path))
-                {
-                    Directory.Delete(path);
-                }
+                if (Directory.Exists(path)) Directory.Delete(path);
 
                 return true;
             }
@@ -61,57 +47,41 @@ namespace Infrastructure.Network.Disk
             }
         }
 
+        public void DeleteFile(string path)
+        {
+            if (string.IsNullOrWhiteSpace(path)) return;
+
+            if (File.Exists(path)) File.Delete(path);
+        }
+
+        public bool DirectoryExists(string path)
+        {
+            return Directory.Exists(path);
+        }
+
         public IReadOnlyCollection<string> GetFiles(string path, string fileMask)
         {
-            if (string.IsNullOrWhiteSpace(path))
-            {
-                return new List<string>();
-            }
+            if (string.IsNullOrWhiteSpace(path)) return new List<string>();
 
             var invalidCharacters = Path.GetInvalidPathChars();
-            if (path.ToCharArray().Any(pa => invalidCharacters.Contains(pa)))
-            {
-                throw new ArgumentException();
-            }
+            if (path.ToCharArray().Any(pa => invalidCharacters.Contains(pa))) throw new ArgumentException();
 
-            if (!Directory.Exists(path))
-            {
-                return new List<string>();
-            }
+            if (!Directory.Exists(path)) return new List<string>();
 
-            if (!string.IsNullOrWhiteSpace(fileMask))
-            {
-                return Directory.GetFiles(path, fileMask);
-            }
-            else
-            {
-                return Directory.GetFiles(path);
-            }
+            if (!string.IsNullOrWhiteSpace(fileMask)) return Directory.GetFiles(path, fileMask);
+            return Directory.GetFiles(path);
         }
 
         public void Move(string originPath, string destinationPath)
         {
-            if (string.IsNullOrWhiteSpace(originPath)
-                || string.IsNullOrWhiteSpace(destinationPath))
-            {
-                return;
-            }
+            if (string.IsNullOrWhiteSpace(originPath) || string.IsNullOrWhiteSpace(destinationPath)) return;
 
             var invalidCharacters = Path.GetInvalidPathChars();
-            if (originPath.ToCharArray().Any(pa => invalidCharacters.Contains(pa)))
-            {
-                throw new ArgumentException();
-            }
+            if (originPath.ToCharArray().Any(pa => invalidCharacters.Contains(pa))) throw new ArgumentException();
 
-            if (destinationPath.ToCharArray().Any(pa => invalidCharacters.Contains(pa)))
-            {
-                throw new ArgumentException();
-            }
+            if (destinationPath.ToCharArray().Any(pa => invalidCharacters.Contains(pa))) throw new ArgumentException();
 
-            if (!File.Exists(originPath))
-            {
-                return;
-            }
+            if (!File.Exists(originPath)) return;
 
             if (File.Exists(destinationPath))
             {
@@ -121,24 +91,6 @@ namespace Infrastructure.Network.Disk
             }
 
             File.Move(originPath, destinationPath);
-        }
-
-        public void DeleteFile(string path)
-        {
-            if (string.IsNullOrWhiteSpace(path))
-            {
-                return;
-            }
-
-            if (File.Exists(path))
-            {
-                File.Delete(path);
-            }
-        }
-        
-        public bool DirectoryExists(string path)
-        {
-            return Directory.Exists(path);
         }
     }
 }
